ui: move date helpers into /helpers
Moves tests out of dedicated /tests folder following #272
