+++ ui/src/components/DateTimePicker.vue
@@ -62,7 +62,7 @@
 import Vue from 'vue'
 import Component from 'vue-class-component'
 import { Prop, Watch } from 'vue-property-decorator'
-import { getISODateString, getISOTimeString } from './date-helpers'
+import { getISODateString, getISOTimeString } from '../helpers/date-helpers'
 
 /**
  * DateTimePicker takes a Date via v-model as input and lets the user select it.
